Simple change for the tutorial to flow
In the section where you are creating the defimpl Size, for X.... the order of creating them and the subsequent calling of them are out of order... not a huge deal but had a small confusion as I was creating each one by hand and testing them. May be of benefit to others in the same scenario.

The fix just changed the order of the defimpl creation to match running of testing them.

Thanks for the great documentation!